requirements.txt
adafruit-circuitpython-sht31d
RPi.GPIO
